refactor(be): introduce xnvme_platform for per-OS backend selection
Separate platform-specific code from runtime backend selection by
introducing xnvme_platform struct with function pointers for dev_open
and enumerate operations.

This refactoring prepares for:

- Centralized reference counting for userspace backends (spdk, vfio)
- Cleaner separation between OS abstraction and I/O path selection

Architecture:

- xnvme_platform: OS-specific, compile-time selected via #ifdef
- platform->backends[]: Runtime-selectable backend array per platform
- Shared implementations iterate backends and configure mixins
