+++ docs/backends/linux/xnvme_info_zoned.out
@@ -17,8 +17,8 @@ xnvme_dev:
     attr: {name: 'linux'}
   xnvme_opts:
     be: 'linux'
-    mem: 'FIX-ID-VS-MIXIN-NAME'
-    dev: 'FIX-ID-VS-MIXIN-NAME'
+    mem: 'posix'
+    dev: 'linux'
     admin: 'block'
     sync: 'block'
     async: 'emu'
