+++ docs/backends/windows/xnvme_win_info_default.out
@@ -13,7 +13,7 @@ xnvme_dev:
   xnvme_opts:
     be: 'windows'
     mem: 'windows'
-    dev: 'FIX-ID-VS-MIXIN-NAME'
+    dev: 'windows'
     admin: 'nvme'
     sync: 'nvme'
     async: 'iocp'
