+++ docs/backends/windows/xnvme_win_info_fs.out
@@ -13,7 +13,7 @@ xnvme_dev:
   xnvme_opts:
     be: 'windows'
     mem: 'windows'
-    dev: 'FIX-ID-VS-MIXIN-NAME'
+    dev: 'windows'
     admin: 'file'
     sync: 'file'
     async: 'iocp'
