+++ docs/getting_started/build_example_windows.out
@@ -14,7 +14,7 @@ xnvme_dev:
   xnvme_opts:
     be: 'windows'
     mem: 'windows'
-    dev: 'FIX-ID-VS-MIXIN-NAME'
+    dev: 'windows'
     admin: 'nvme'
     sync: 'nvme'
     async: 'iocp'
