+++ include/libxnvme_be.h
@@ -43,7 +43,7 @@ int
 xnvme_be_attr_pr(const struct xnvme_be_attr *attr, int opts);
 
 /**
- * Prints all the backends in the backend registry to the given output stream
+ * Prints platform name and available drivers to the given output stream
  *
  * @param stream output stream used for printing
  * @param opts printer options, see ::xnvme_pr
@@ -54,9 +54,8 @@ int
 xnvme_be_registry_fpr(FILE *stream, enum xnvme_pr opts);
 
 /**
- * Prints all the backends in the backend registry to stdout
+ * Prints platform name and available drivers to stdout
  *
- * @param attr Pointer to the ::xnvme_be_attr to print
  * @param opts printer options, see ::xnvme_pr
  *
  * @return On success, the number of characters printed is returned.
