+++ lib/meson.build
@@ -13,6 +13,11 @@ xnvmelib_source = [
   'xnvme_be_cbi_mem_posix.c',
   'xnvme_be_cref.c',
   'xnvme_be_cbi_sync_psync.c',
+  'xnvme_platform.c',
+  'xnvme_platform_linux.c',
+  'xnvme_platform_freebsd.c',
+  'xnvme_platform_macos.c',
+  'xnvme_platform_windows.c',
   'xnvme_be_fbsd.c',
   'xnvme_be_fbsd_async.c',
   'xnvme_be_fbsd_dev.c',
