+++ lib/xnvme_be_linux_nvme.c
@@ -8,7 +8,7 @@
 #include <libxnvme.h>
 #include <xnvme_be.h>
 #include <xnvme_be_nosys.h>
-#ifdef XNVME_BE_LINUX_ENABLED
+#ifdef XNVME_PLATFORM_LINUX_ENABLED
 #include <errno.h>
 #include <fcntl.h>
 #include <linux/nvme_ioctl.h>
@@ -396,7 +396,7 @@ xnvme_be_linux_nvme_supported(struct xnvme_dev *dev, uint32_t XNVME_UNUSED(opts)
 
 struct xnvme_be_sync g_xnvme_be_linux_sync_nvme = {
 	.id = "nvme",
-#ifdef XNVME_BE_LINUX_ENABLED
+#ifdef XNVME_PLATFORM_LINUX_ENABLED
 	.cmd_io = xnvme_be_linux_nvme_cmd_io,
 	.cmd_iov = xnvme_be_linux_nvme_cmd_iov,
 #else
@@ -407,7 +407,7 @@ struct xnvme_be_sync g_xnvme_be_linux_sync_nvme = {
 
 struct xnvme_be_admin g_xnvme_be_linux_admin_nvme = {
 	.id = "nvme",
-#ifdef XNVME_BE_LINUX_ENABLED
+#ifdef XNVME_PLATFORM_LINUX_ENABLED
 	.cmd_admin = xnvme_be_linux_nvme_cmd_admin,
 	.cmd_pseudo = xnvme_be_linux_nvme_cmd_pseudo,
 #else
