+++ lib/xnvme_libconf_entries.c
@@ -33,8 +33,8 @@ static const char *g_libconf_strings[] = {
 #ifdef XNVME_BE_RAMDISK_ENABLED
 	"conf: XNVME_BE_RAMDISK_ENABLED",
 #endif
-#ifdef XNVME_BE_WINDOWS_ENABLED
-	"conf: XNVME_BE_WINDOWS_ENABLED",
+#ifdef XNVME_PLATFORM_WINDOWS_ENABLED
+	"conf: XNVME_PLATFORM_WINDOWS_ENABLED",
 #endif
 #ifdef XNVME_BE_WINDOWS_FS_ENABLED
 	"conf: XNVME_BE_WINDOWS_FS_ENABLED",
@@ -42,8 +42,8 @@ static const char *g_libconf_strings[] = {
 #ifdef XNVME_BE_WINDOWS_ASYNC_ENABLED
 	"conf: XNVME_BE_WINDOWS_ASYNC_ENABLED",
 #endif
-#ifdef XNVME_BE_LINUX_ENABLED
-	"conf: XNVME_BE_LINUX_ENABLED",
+#ifdef XNVME_PLATFORM_LINUX_ENABLED
+	"conf: XNVME_PLATFORM_LINUX_ENABLED",
 #endif
 #ifdef XNVME_BE_LINUX_BLOCK_ENABLED
 	"conf: XNVME_BE_LINUX_BLOCK_ENABLED",
@@ -78,11 +78,11 @@ static const char *g_libconf_strings[] = {
 #ifdef XNVME_BE_SPDK_TRANSPORT_FC_ENABLED
 	"conf: XNVME_BE_SPDK_TRANSPORT_FC_ENABLED",
 #endif
-#ifdef XNVME_BE_MACOS_ENABLED
-	"conf: XNVME_BE_MACOS_ENABLED",
+#ifdef XNVME_PLATFORM_MACOS_ENABLED
+	"conf: XNVME_PLATFORM_MACOS_ENABLED",
 #endif
-#ifdef XNVME_BE_FBSD_ENABLED
-	"conf: XNVME_BE_FBSD_ENABLED",
+#ifdef XNVME_PLATFORM_FREEBSD_ENABLED
+	"conf: XNVME_PLATFORM_FREEBSD_ENABLED",
 #endif
 
 #ifdef NVME_IOCTL_IO64_CMD
